refactor(test): shift common testlib items into a 'common' package
this means they have their own separate directory with no collisions
in other modules that consume them
